+++ src/Symfony/Component/Uid/BinaryUtil.php
@@ -119,10 +119,10 @@ class BinaryUtil
     /**
      * @param string $time Count of 100-nanosecond intervals since the UUID epoch 1582-10-15 00:00:00 in hexadecimal
      */
-    public static function timeToFloat(string $time): float
+    public static function timeToDateTime(string $time): \DateTimeImmutable
     {
         if (\PHP_INT_SIZE >= 8) {
-            $time = hexdec($time) - self::TIME_OFFSET_INT;
+            $time = (string) (hexdec($time) - self::TIME_OFFSET_INT);
         } else {
             $time = str_pad(hex2bin($time), 8, "\0", \STR_PAD_LEFT);
 
@@ -136,6 +136,10 @@ class BinaryUtil
             }
         }
 
-        return $time / 10000000;
+        if (9 > \strlen($time)) {
+            $time = '-' === $time[0] ? '-'.str_pad(substr($time, 1), 8, '0', \STR_PAD_LEFT) : str_pad($time, 8, '0', \STR_PAD_LEFT);
+        }
+
+        return \DateTimeImmutable::createFromFormat('U.u?', substr_replace($time, '.', -7, 0));
     }
 }